Finished 64-bit segmented bfs functionality
The code is currently UNTESTED. Most of the
error-prone logic to handle segmented ints is
at the swig layer, but some is also at the python
layer. For example, GetDistCpp64.py infers which
segment it is responsible for by looking at the
node for which its Range starts.

In order for this to work properly, all node ranges
must divide evenly into the segment size, which is
equal to (1 << seg_bits), where seg_bits is specified
in the config file.

The next step is to test the 64 bit app against the
non 64 bit equivalent, with the same random seed, and
ensure that results are the same. After that, I need
to somehow make sure things don't break for non-toy
problems as well, where some of the nodes actually
do have id's greater than INT_MAX.